requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu>=1.7.4
torch>=2.2.0
transformers>=4.35.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy>=1.24.3
pandas>=2.1.3
tiktoken>=0.5.1 